使用DI优化了代码,加入邮件送信功能,加入授权策略,加入任务队列

+++ app/Http/Controllers/Admin/ArticleController.php
@@ -77,7 +77,9 @@ class ArticleController extends Controller
             'body' => 'required',
         ]);
 
-        $article = Article::find($id);
+        $article = Article::findOrFail($id);
+        $this->authorize('update', $article);
+
         $article->title = $request->get('title');
         $article->body = $request->get('body');
 
@@ -94,8 +96,10 @@ class ArticleController extends Controller
      */
     public function destroy($id)
     {
+        $article = Article::findOrFail($id);
+        $this->authorize('delete', $article);
         Comment::where('article_id', '=', $id)->delete();
-        Article::find($id)->delete();
+        $article->delete();
 
         return redirect()->back()->withInput()->withErrors(trans('errors.delete-success'));
     }
